Fix cloudfunction triggers and reconciling old data

+++ functions/lib/index.js
@@ -22,63 +22,61 @@ const config = {
 admin.initializeApp(config);
 exports.getToken = functions.firestore
     .document('tokens/{uid}')
-    .onWrite((change, context) => __awaiter(this, void 0, void 0, function* () {
+    .onCreate((change, context) => __awaiter(this, void 0, void 0, function* () {
     const redirect_uri = 'https://bard.jaydanhoward.com/authenticate/';
-    const oldData = change.before.data();
-    const oldCode = oldData ? change.before.data().auth_code : null;
-    const newCode = change.after.data().auth_code;
+    const code = change.data().auth_code;
     const { id, secret } = yield (yield admin
         .firestore()
         .collection('client')
         .doc('apiKey')
         .get()).data();
     const json = yield (yield node_fetch_1.default('https://accounts.spotify.com/api/token', {
-        body: `grant_type=authorization_code&client_id=${id}&client_secret=${secret}&code=${newCode}&redirect_uri=${redirect_uri}`,
+        body: `grant_type=authorization_code&client_id=${id}&client_secret=${secret}&code=${code}&redirect_uri=${redirect_uri}`,
         headers: {
             'Content-Type': 'application/x-www-form-urlencoded'
         },
         method: 'POST'
     })).json();
     console.log(json);
-    const useNew = oldCode !== newCode;
-    return json.access_token || (!useNew && oldData)
-        ? change.after.ref.set({
-            access_token: useNew ? json.access_token : oldData.access_token,
-            refresh_token: useNew ? json.refresh_token : oldData.refresh_token,
-            auth_code: useNew ? newCode : oldData.auth_code,
-            time: useNew ? Date.now() : oldData.time
+    return json.access_token
+        ? change.ref.set({
+            access_token: json.access_token,
+            refresh_token: json.refresh_token,
+            auth_code: code,
+            time: Date.now()
         })
-        : change.after.ref.delete();
+        : change.ref.delete();
+    return Promise.resolve();
 }));
 exports.refreshToken = functions.firestore
     .document('tokens/{uid}')
     .onWrite((change, context) => __awaiter(this, void 0, void 0, function* () {
-    const oldFetch = change.before.data().refetch;
-    const newFetch = change.after.data().refetch;
-    const refresh_token = change.after.data().refresh_token;
-    const { id, secret } = yield (yield admin
-        .firestore()
-        .collection('client')
-        .doc('apiKey')
-        .get()).data();
-    const json = yield (yield node_fetch_1.default('https://accounts.spotify.com/api/token', {
-        body: `grant_type=refresh_token&client_id=${id}&client_secret=${secret}&refresh_token=${refresh_token}`,
-        headers: {
-            'Content-Type': 'application/x-www-form-urlencoded'
-        },
-        method: 'POST'
-    })).json();
-    return newFetch && newFetch !== oldFetch
-        ? change.after.ref.set({
-            access_token: json.access_token
-                ? json.access_token
-                : change.after.data().access_token,
-            refresh_token: json.refresh_token
-                ? json.refresh_token
-                : refresh_token,
-            refetch: false,
-            time: Date.now()
-        }, { merge: true })
-        : null;
+    const oldFetch = change.before.data() ? change.before.data().refetch : null;
+    const newFetch = change.after.data() ? change.after.data().refetch : null;
+    if (newFetch && newFetch !== oldFetch) {
+        const refresh_token = change.after.data().refresh_token;
+        const { id, secret } = yield (yield admin
+            .firestore()
+            .collection('client')
+            .doc('apiKey')
+            .get()).data();
+        const json = yield (yield node_fetch_1.default('https://accounts.spotify.com/api/token', {
+            body: `grant_type=refresh_token&client_id=${id}&client_secret=${secret}&refresh_token=${refresh_token}`,
+            headers: {
+                'Content-Type': 'application/x-www-form-urlencoded'
+            },
+            method: 'POST'
+        })).json();
+        return json.access_token
+            ? change.after.ref.set({
+                access_token: json.access_token
+                    ? json.access_token
+                    : change.after.data().access_token,
+                refetch: false,
+                time: Date.now()
+            }, { merge: true })
+            : change.after.ref.delete();
+    }
+    return Promise.resolve();
 }));
 //# sourceMappingURL=index.js.map
